test: rename package test files to avoid name collision after transpilation
